+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - 优化 `f-avatar` 组件部分样式的实现，减少代码开销
 - 优化 `f-back-top` 组件部分样式的实现，减少代码开销
 - `f-back-top` 组件 `right` `bottom` 配置项支持 `number` 类型
+- `f-image` 组件 `round` 配置项支持 `number` 类型
 - 优化 `f-back-top` 组件部分 `props` 默认值
 - 优化 `f-badge` 组件样式实现